chore: external 경로 수정
