+++ src/main/java/com/nhnacademy/illuwa/common/external/user/dto/MemberSavePointDto.java
@@ -1,19 +1,14 @@
-package com.nhnacademy.illuwa.domain.order.external.member;
+package com.nhnacademy.illuwa.common.external.user.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 
 import java.math.BigDecimal;
 
-@Setter
 @Getter
-@AllArgsConstructor
 @NoArgsConstructor
-public class TotalRequest {
-
-    private Long memberId;
+@AllArgsConstructor
+public class MemberSavePointDto {
     private BigDecimal totalPrice;
-
 }
