+++ src/main/java/com/nhnacademy/illuwa/common/external/user/dto/MemberUsedPointDto.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.illuwa.domain.order.external.member.dto;
+package com.nhnacademy.illuwa.common.external.user.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -9,7 +9,6 @@ import java.math.BigDecimal;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class MemberPointDto {
-
-    private BigDecimal point;
+public class MemberUsedPointDto {
+    private BigDecimal usedPoint;
 }
