+++ src/main/java/com/nhnacademy/illuwa/common/external/user/dto/TotalRequest.java
@@ -1,14 +1,19 @@
-package com.nhnacademy.illuwa.domain.order.external.member.dto;
+package com.nhnacademy.illuwa.common.external.user.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import lombok.Setter;
 
 import java.math.BigDecimal;
 
+@Setter
 @Getter
-@NoArgsConstructor
 @AllArgsConstructor
-public class MemberSavePointDto {
+@NoArgsConstructor
+public class TotalRequest {
+
+    private Long memberId;
     private BigDecimal totalPrice;
+
 }
