+++ src/main/java/com/nhnacademy/illuwa/domain/order/factory/OrderFactory.java
@@ -10,8 +10,8 @@ import com.nhnacademy.illuwa.domain.order.entity.Packaging;
 import com.nhnacademy.illuwa.domain.order.entity.ShippingPolicy;
 import com.nhnacademy.illuwa.domain.order.entity.types.OrderStatus;
 import com.nhnacademy.illuwa.domain.order.exception.common.NotFoundException;
-import com.nhnacademy.illuwa.domain.order.external.book.BookPriceApiClient;
-import com.nhnacademy.illuwa.domain.order.external.book.BookPriceDto;
+import com.nhnacademy.illuwa.common.external.book.BookPriceApiClient;
+import com.nhnacademy.illuwa.common.external.product.dto.BookPriceDto;
 import com.nhnacademy.illuwa.domain.order.repository.OrderRepository;
 import com.nhnacademy.illuwa.domain.order.repository.PackagingRepository;
 import com.nhnacademy.illuwa.domain.order.repository.ShippingPolicyRepository;
