+++ src/main/java/com/nhnacademy/illuwa/domain/order/service/impl/OrderServiceImpl.java
@@ -8,12 +8,12 @@ import com.nhnacademy.illuwa.domain.order.entity.Order;
 import com.nhnacademy.illuwa.domain.order.entity.types.OrderStatus;
 import com.nhnacademy.illuwa.domain.order.exception.common.NotFoundException;
 import com.nhnacademy.illuwa.domain.order.exception.common.NotFoundStringException;
-import com.nhnacademy.illuwa.domain.order.external.cart.CartApiClient;
-import com.nhnacademy.illuwa.domain.order.external.cart.dto.CartOrderItemDto;
-import com.nhnacademy.illuwa.domain.order.external.cart.dto.CreateOrderFromCartRequest;
-import com.nhnacademy.illuwa.domain.order.external.member.MemberAddressApiClient;
-import com.nhnacademy.illuwa.domain.order.external.member.MemberPointApiClient;
-import com.nhnacademy.illuwa.domain.order.external.member.dto.MemberAddressDto;
+import com.nhnacademy.illuwa.common.external.cart.CartApiClient;
+import com.nhnacademy.illuwa.common.external.product.dto.CartOrderItemDto;
+import com.nhnacademy.illuwa.common.external.product.dto.CreateOrderFromCartRequest;
+import com.nhnacademy.illuwa.common.external.member.MemberAddressApiClient;
+import com.nhnacademy.illuwa.common.external.member.MemberPointApiClient;
+import com.nhnacademy.illuwa.common.external.user.dto.MemberAddressDto;
 import com.nhnacademy.illuwa.domain.order.factory.OrderFactory;
 import com.nhnacademy.illuwa.domain.order.repository.OrderRepository;
 import com.nhnacademy.illuwa.domain.order.service.OrderService;
